requirements.txt
msal
httpx
pydantic
python-dotenv
anthropic
typer
